Reuse safe Frame source copying
